[vegastrike] improve c++11 and clang / gcc>6 compatibility
It nows builds with apple clang 11, clang 12, gcc 6 to 11

* removed a lot of pointers signed comparison (eg ptr > 0)

* removed some functions returning char as char*

* use global vegastrike version in setup

* detect clang libc++ or gnu libstdc++ version (base_computer.cpp)

* some static_cast double to float needed (al_sound,

* size_t/unisgned int , void */char* casts

* remove some static class method calls from instance (unit_collide,hashtable_3d)

* remove some strings literals with MACROS without space ("abc"VERSION)

* src/cmd/collection.h -> use the STL collection (new)
  c++11 error with active_huge.size()

* ...

+++ vegastrike/src/aldrv/al_sound.cpp
@@ -707,8 +707,8 @@ void AUDDeleteSound (int sound, bool music){
 void AUDAdjustSound (const int sound, const QVector &pos, const Vector &vel){
 #ifdef HAVE_AL
   if (sound>=0&&sound<(int)sounds.size()) {
-    float p []= {scalepos*pos.i,scalepos*pos.j,scalepos*pos.k};
-    float v []= {scalevel*vel.i,scalevel*vel.j,scalevel*vel.k};
+    float p []= {static_cast<float>(scalepos*pos.i),static_cast<float>(scalepos*pos.j),static_cast<float>(scalepos*pos.k)};
+    float v []= {static_cast<float>(scalevel*vel.i),static_cast<float>(scalevel*vel.j),static_cast<float>(scalevel*vel.k)};
     sounds[sound].pos = pos.Cast();
 	sounds[sound].vel=vel;
 	if (usepositional&&sounds[sound].source) {
