+++ vegastrike/src/vs_random.h
@@ -1,10 +1,10 @@
-/* 
+/*
    A C-program for MT19937, with initialization improved 2002/1/26.
    Coded by Takuji Nishimura and Makoto Matsumoto.
-   Before using, initialize the state by using init_genrand(seed)  
+   Before using, initialize the state by using init_genrand(seed)
    or init_by_array(init_key, key_length).
    Copyright (C) 1997 - 2002, Makoto Matsumoto and Takuji Nishimura,
-   All rights reserved.                          
+   All rights reserved.
    Redistribution and use in source and binary forms, with or without
    modification, are permitted provided that the following conditions
    are met:
@@ -13,8 +13,8 @@
      2. Redistributions in binary form must reproduce the above copyright
         notice, this list of conditions and the following disclaimer in the
         documentation and/or other materials provided with the distribution.
-     3. The names of its contributors may not be used to endorse or promote 
-        products derived from this software without specific prior written 
+     3. The names of its contributors may not be used to endorse or promote
+        products derived from this software without specific prior written
         permission.
    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
    "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
@@ -35,9 +35,9 @@
 /*
 #define N 624
 #define M 397
-#define MATRIX_A 0x9908b0dfUL   // constant vector a 
-#define UPPER_MASK 0x80000000UL // most significant w-r bits 
-#define LOWER_MASK 0x7fffffffUL // least significant r bits 
+#define MATRIX_A 0x9908b0dfUL   // constant vector a
+#define UPPER_MASK 0x80000000UL // most significant w-r bits
+#define LOWER_MASK 0x7fffffffUL // least significant r bits
 */
 
 #define VS_RAND_MAX 0x7fffffffUL
@@ -70,8 +70,8 @@ public:
   void init_genrand(unsigned int s) {
 		mt[0]= s & 0xffffffffUL;
 		for (mti=1; mti<N(); mti++) {
-			mt[mti] = 
-				(1812433253UL * (mt[mti-1] ^ (mt[mti-1] >> 30)) + mti); 
+			mt[mti] =
+				(1812433253UL * (mt[mti-1] ^ (mt[mti-1] >> 30)) + mti);
 			/* See Knuth TAOCP Vol2. 3	rd Ed. P.106 for multiplier. */
 			/* In the previous versions, MSB	s of the seed affect   */
 			/* only MSBs of the array mt[].         	               */
@@ -103,12 +103,12 @@ public:
         i++;
         if (i>=N()) { mt[0] = mt[N()-1]; i=1; }
     }
-    mt[0] = 0x80000000UL; /* MSB is 1; assuring non-zero initial array */ 
+    mt[0] = 0x80000000UL; /* MSB is 1; assuring non-zero initial array */
   }
 /* generates a random number on [0,0xffffffff]-interval */
   unsigned int genrand_int32(void) {
     unsigned int y;
-    static unsigned int mag01[2]={0x0UL, MATRIX_A()};
+    static unsigned int mag01[2]={0x0UL, static_cast<unsigned int>((MATRIX_A)())};
     /* mag01[x] = x * MATRIX_A  for x=0,1 */
     if (mti >= N()) { /* generate N words at one time */
         int kk;
@@ -126,7 +126,7 @@ public:
         mt[N()-1] = mt[M()-1] ^ (y >> 1) ^ mag01[y & 0x1UL];
         mti = 0;
     }
-  
+
     y = mt[mti++];
     /* Tempering */
     y ^= (y >> 11);
@@ -144,13 +144,13 @@ public:
   }
 /* generates a random number on [0,1]-real-interval */
   double genrand_real1(void) {
-    return genrand_int32()*(1.0/4294967295.0); 
-    /* divided by 2^32-1 */ 
+    return genrand_int32()*(1.0/4294967295.0);
+    /* divided by 2^32-1 */
   }
 /* generates a random number on [0,1)-real-interval */
   double genrand_real2(void)
   {
-    return genrand_int32()*(1.0/4294967296.0); 
+    return genrand_int32()*(1.0/4294967296.0);
     /* divided by 2^32 */
   }
   double uniformInc (double min, double max) {
@@ -159,17 +159,17 @@ public:
   double uniformExc(double min, double max) {
 	  return genrand_real2()*(max-min)+min;
   }
-	
+
 /* generates a random number on (0,1)-real-interval */
   double genrand_real3(void) {
-    return (((double)genrand_int32()) + 0.5)*(1.0/4294967296.0); 
+    return (((double)genrand_int32()) + 0.5)*(1.0/4294967296.0);
     /* divided by 2^32 */
   }
 /* generates a random number on [0,1) with 53-bit resolution*/
-  double genrand_res53(void) { 
-    unsigned int a=genrand_int32()>>5, b=genrand_int32()>>6; 
-    return(a*67108864.0+b)*(1.0/9007199254740992.0); 
-  } 
+  double genrand_res53(void) {
+    unsigned int a=genrand_int32()>>5, b=genrand_int32()>>6;
+    return(a*67108864.0+b)*(1.0/9007199254740992.0);
+  }
 /* These real versions are due to Isaku Wada, 2002/01/09 added */
 };
 extern VSRandom vsrandom;
